fix : 타입 수정

+++ src/main/java/com/example/ssg_tab/domain/contents/repository/ContentsRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.Collection;
 import java.util.List;
 
-public interface ContentsRepository extends JpaRepository<Contents, Integer> {
+public interface ContentsRepository extends JpaRepository<Contents, Long> {
     @Query("select c.sourceUrl from Contents c where c.sourceUrl in :urls")
     List<String> findExistingUrls(Collection<String> urls); // 원본 URL이 같은 것 찾기
